refactor(csv-stringify): cast types isolation
