+++ packages/csv-stringify/test/option.cast.ts
@@ -20,96 +20,6 @@ describe("Option `cast`", function () {
   });
 
   describe("udf", function () {
-    it("handle string formatter", function (next) {
-      stringify(
-        [
-          {
-            value: "ok",
-          },
-        ],
-        { cast: { string: () => "X" } },
-        (err, data) => {
-          if (!err) data.should.eql("X\n");
-          next(err);
-        },
-      );
-    });
-
-    it("handle boolean formatter", function (next) {
-      stringify(
-        [
-          {
-            value: true,
-          },
-        ],
-        { cast: { boolean: () => "X" } },
-        (err, data) => {
-          if (!err) data.should.eql("X\n");
-          next(err);
-        },
-      );
-    });
-
-    it("handle date formatter", function (next) {
-      stringify(
-        [
-          {
-            value: new Date(),
-          },
-        ],
-        { cast: { date: () => "X" } },
-        (err, data) => {
-          if (!err) data.should.eql("X\n");
-          next(err);
-        },
-      );
-    });
-
-    it("handle number formatter", function (next) {
-      stringify(
-        [
-          {
-            value: 3.14,
-          },
-        ],
-        { cast: { number: (value) => "" + value * 2 } },
-        (err, data) => {
-          if (!err) data.should.eql("6.28\n");
-          next(err);
-        },
-      );
-    });
-
-    it("handle bigint formatter", function (next) {
-      stringify(
-        [
-          {
-            value: BigInt(9007199254740991),
-          },
-        ],
-        { cast: { bigint: (value) => "" + value / BigInt(2) } },
-        (err, data) => {
-          if (!err) data.should.eql("4503599627370495\n");
-          next(err);
-        },
-      );
-    });
-
-    it("handle object formatter", function (next) {
-      stringify(
-        [
-          {
-            value: { a: 1 },
-          },
-        ],
-        { cast: { object: () => "X" } },
-        (err, data) => {
-          if (!err) data.should.eql("X\n");
-          next(err);
-        },
-      );
-    });
-
     it("catch error", function (next) {
       stringify(
         [
